Check formatting changes only in formatter
Move the logic to check if the mapping changes the formatting of the
matched text and what to do accordingly to the formatter (formatting.go)
in order to fix the latest test for this new feature. The refactoring
allows us to perform th `maintainWhitespace` function without executing
the `maintainAllCaps` and `maintainCapitalization` functions.

Consequently the (admittedly out-of-place) strings.ToLower function in
mapping.getReplacement has been moved to the formatter as well.