Include UTF-8  check in Mapping.Match
Update the logic behind Mapping.Match to ignore any mapping containing
non-UTF8 characters. In that case, a Warning is outputted. There is also
now a Notes section in the documentation mappings which informs users
that this is the case.

In addition, the Mapping fuzz function no longer ignores test cases with
non-UTF8 characters.