Expand unit test suite for Mapping.Match
Add two new unit tests for Mapping.Match that test for the recently
discovered issues in the `getAllMatches` function w.r.t to non-UTF8
characters and invalid Regular Expressions in mapping values.

Even though the tests or not super comprehensive (when compared to the
fuzzing used to discover them), there is a test for just one non-UTF8
character and a test for one invalid regular expression, they should be
able to catch any problems regarding these issues in the future when
running the standard test suite (without having to resort to fuzzing).