Test and fix crashes due to mappings with invalid Regular Expression syntax (#96)
* Create fuzzing function for wordmaps.Mapping

Specifically, a fuzzing function that fuzzes the `getAllMatches`
function of the mapping.go file in the wordmaps package.

This function uses a regular expression that is build using user input.
Therefore, the function is vulnerable to attacks and can easily be used
to crash wordrow.

* Update Mapping fuzzer to ignore non-UTF8 queries

Update the new fuzzing function for wordmaps.Mapping to temporarily
ignore any queries (i.e. `substr`s) that contain a character that is not
UTF-8. This reduces the number of crashes generated greatly so that
other issues can be found more easily.

* Improve output when running `make fuzz`

Instead of seeing the raw commands that are executed, running make fuzz
will now output:

```
BUILDING FUZZING BINARY FOR internal/wordmaps...
```

while running go-fuzz-build, and after that

```
FUZZING internal/wordmaps::FuzzMapping...
```

when it starts fuzzing, followed by the output of `go-fuzz`. Resulting
in something that looks like:

```
$ make fuzz PKG=internal/wordmaps FUNC=FuzzMapping
BUILDING FUZZING BINARY FOR internal/wordmaps...
FUZZING internal/wordmaps::FuzzMapping...
2020/07/22 00:00:00 workers: 4, corpus: 116 (0s ago), crashers: 135, restarts: 1/0, execs: 0 (0/sec), cover: 0, uptime: 3s
....
```

* Escape all special characters in getAllMatches

Update the getAllMatches function for the mapping package so that the
query strings (parameters `substr`) are escape. To achieve this the
QuoteMeta function from the regexp package [1] was used. This function
will automatically escape all special characters for regular epxression
in the provided string. Unfortuantely, this doesn't work very  nicely
with the mapping syntax prefix/suffix notation escape technique. For now
this is resolved by unescaping hyphens and backslashes before calling
QuoteMeta.

--
1. https://golang.org/pkg/regexp/#QuoteMeta

* Update mapping_fuzz to fuzz the Mapping struct

... instead of a function used by the Mapping struct.

* Include UTF-8  check in Mapping.Match

Update the logic behind Mapping.Match to ignore any mapping containing
non-UTF8 characters. In that case, a Warning is outputted. There is also
now a Notes section in the documentation mappings which informs users
that this is the case.

In addition, the Mapping fuzz function no longer ignores test cases with
non-UTF8 characters.

* Expand unit test suite for Mapping.Match

Add two new unit tests for Mapping.Match that test for the recently
discovered issues in the `getAllMatches` function w.r.t to non-UTF8
characters and invalid Regular Expressions in mapping values.

Even though the tests or not super comprehensive (when compared to the
fuzzing used to discover them), there is a test for just one non-UTF8
character and a test for one invalid regular expression, they should be
able to catch any problems regarding these issues in the future when
running the standard test suite (without having to resort to fuzzing).

* Update CHANGELOG