Define mapping to change capitalization of a word/phrase (#98)
* Add test to "replace" and change capitalization

This test tests the feature request of issue 87. This issue specifies
that it should be possible to fix capitalization of words/phrases using
wordrow. This test assumes the proposed implementation, where a mapping
that has two of the same values except for their capitalization shall be
interpreted as having to change the capitalization of words explicitly.

* Proof of concept implementation of changing capitalisation

Proof of concept implementation of a replacement where the capitalisation
changes based on the mapping rather than the document context. To be
able to do this, the WordMap struct had to be updated so that it no longer
stores all values in all lowercase. However, the replace package assumed
this to be the case. Hence, this is now implemented by providing the
replacement in all-lowercase. Then, the replaceOne function will check if
this mapping is a "changes formatting" mapping and ignore the formatting-
maintained replacement and use the exact To value instead.

* More test cases for replace change capitalization

add additional test cases for the new replace feature where
capitalization is explicitly changed, namely:

- Change string to all lowercase
- Change to all caps
- Change capitalization with multiple words in a phrase

* Add test for mapping.ChangesFormatting

* Test capitalization changing with newline inserted

Add a test to see if the capitalization changing mappings are handled
correctly if there is different whitespace in between the words of the
mapping in the source text.

* Check formatting changes only in formatter

Move the logic to check if the mapping changes the formatting of the
matched text and what to do accordingly to the formatter (formatting.go)
in order to fix the latest test for this new feature. The refactoring
allows us to perform the `maintainWhitespace` function without executing
the `maintainAllCaps` and `maintainCapitalization` functions.

Consequently the (admittedly out-of-place) strings.ToLower function in
mapping.getReplacement has been moved to the formatter as well.

* Remove unused mapping.ChangesFormatting

* Update CHANGELOG

* Update documentation regarding capitalisation

Update the documentation of mapping files with regards to capitalisation
to include the fact that mappings where the values are textually the
same will result in the capitalisation of the to value being maintained.