Fix bug for mapfile args without file extension
Update the parseMapFileArgument function in the main package so that it
can handle an argument of a file without a file extension with an
explicit format. The function documentation has been updated to be more
precise in this regard as well.

The bug was caused by the fact that the program looked for the explicit
format in the string representing the file extension. For files without
an extension this string is empty and hence now explicit format could be
found. Hence, to resolve this the explicit format is now searched for in
the full argument string.