Support and test multiple equals in argument
Add support for arguments of the form "--arg=value" to contain more = in
the value. Previously this form was only supported if there is exactly
one = in the argument. Now it can have many equals sign an the string
before the first = is interpreted as the option and all the rest
together as the value for that option.

This has also been tested with one simple example.