Add support for arguments with equals sign (#102)
* Add test for arguments with equals

Add a test that verifies if a program argument like "--map=foo,bar",
with an equals sign to separate the option from the value,  is handled
by the program.

* POC implementation of argument with equals

Add a basic Proof of Concept (POC) implementation of that supports
parsing CLI arguments to contain an equals sign to separate an option
from its value.

* Verify that no value is set upon invalid argument when using the equals
notation for arguments.

* Refactor parsing of argument as option

* Update CHANGELOG

* Support and test multiple equals in argument

Add support for arguments of the form "--arg=value" to contain more = in
the value. Previously this form was only supported if there is exactly
one = in the argument. Now it can have many equals sign an the string
before the first = is interpreted as the option and all the rest
together as the value for that option.

This has also been tested with one simple example.