Run Security Workflow on pushes
Add support for the GitHub Actions Security Workflow to run on pushes
(instead of only on Pull Requests). This Workflow will not be triggered
for pushes to all branches. Most importantly, it runs the Security
Workflow on the master branch. Additionally, it also runs on any branch
whose name starts with "sec" (short for "security") as that is an
indication that the branch is concerned with security and therefore
would be interested in (automated) security feedback.