Add errcheck static analysis tool
Run errcheck [1] as a static analysis tool upon `make analysis`. This
tool verifies that all errors are accounter for. The errcheck_
excludes.txt file (name based on the errcheck documentation) is used to
specify function calls whose errors can be ignored safely.

Currently, three function calls are ignored. The first is that of a file
handle being closed, which only errors when already closed (which is not
a problem in this project). The others are related to strings.Builder,
whose errors can generally be ignored safely.

--
1. https://github.com/kisielk/errcheck