Add support for many-to-one mapping syntax on the CLI (#118)
* Use wordmaps.AddFile for CLI defined mpapings

Refactor the `processInlineMapping` function in package main to just
use `wordmaps.AddFile` with the CSV format to process mappings defined
on the CLI (using --map or -m). This replaces a custom and simplified
CSV parser that was there before.

This approach has the advantage that CLI specified mappings will
benefit from any improvements/bugs fixed in the CSV parser. As opposed
to having to improve/fix it in two places. The main disadvantage is that
the full CSV parser in package wordmaps is  bit overkill for the CLI
defined mappings.

One notable alternative implementation would be to accumulate all CLI
defined mappings into a long string, and run wordmap.AddFile only once.
This would (arguably) make the code a bit more sensible, however, then
any error that occurs prevents later mappings from being used. Though
this may be a valid way to handle such a case, it is not compatible with
the current behaviour of the program so I decided to leave this as is.

* Update CHANGELOG