Add nakedret, unimport, gochecknoinits as linters
Add a couple of linting tools (differentiated from "analysis" in that
they focus more, arguably, on stylistic preferences). In particular:

- nakedret [1], which is currently used to completely disallow naked
  return statements.
- unimport [2], which disallows import renaming unless it is necessary
  to avoid naming conflicts.
- gochecknoinits [3], which is used to disallow `init()` functions
  (which may cause unexpected side effects).

One other tool that has been considered but not added is
gochecknoglobals [4] which would disallow Package-scope globals. This
would further reduce side-effects. I decided not too add this as it
seems a bit too restrictive for my taste, and would require too much
refactoring at this stage of development.

--
1. https://github.com/alexkohler/nakedret
2. https://github.com/alexkohler/unimport
3. https://github.com/leighmcculloch/gochecknoinits
4. https://github.com/leighmcculloch/gochecknoglobals