Continue on non-blocking errors (#112)
* Continue replacement after error

Update the `processInputFiles` function to continue with other files if
an error occurred in one file (that is not the last file).

* Return all errors when resolving globs

* Return all errors when processing map args

* Add strict flag to CLI interface

Update the CLI interface to include a new "Strict" mode flag. When the
"strict" flag is used, the program will halt when a non-blocking error
occurs. If the "strict" flag is not used this is not the case (to be
implemented).

* Implement basic functionality of strict mode

I.e. return upon errors produced when trying the get the wordmap or
when trying to resolve globs for input files.

* Refactor duplication of file iteration in package main

* Update CHANGELOG