Add looppointer static analysis tool
Run looppointer [1] as a static analysis tool upon `make analysis`. This
tool finds pointers for loops variables which are a common cause of bugs.

I opted for looppointer [1] over exportloopref [2] as it will report
more issues. Since there are currently no issues, I think this is the best
approach to avoid introducing any bugs in the future.

--
1. https://github.com/kyoh86/looppointer
2. https://github.com/kyoh86/exportloopref