Add more static analysis (#113)
* Add errcheck static analysis tool

Run errcheck [1] as a static analysis tool upon `make analysis`. This
tool verifies that all errors are accounted for. The .errcheckrc.txt file
is used to specify function calls whose errors can be ignored safely.

Currently, three function calls are ignored. The first is that of a file
handle being closed, which only errors when already closed (which is not
a problem in this project). The others are related to strings.Builder,
whose errors can generally be ignored safely.

--
1. https://github.com/kisielk/errcheck


* Update CLI output for make analysis

Update the output on the CLI when running make analysis. The commands
themselves are no longer shown. Instead, the text describing the command
is shown with three dots suffixed to indicate the computer is busy.


* Add ineffassign static analysis tool

Run ineffassign [1] as a static analysis tool upon `make analysis`. This
tool verifies that all variable assignment actually have an effect. Even
though it is not perfect (it may not catch some ineffective
assignments), it won't (shouldn't) report any false positives.

--
1.  https://github.com/gordonklaus/ineffassign


* Add deadcode static analysis tool

Run deadcode [1] as a static analysis tool upon `make analysis`. This
tool checks the code base for any dead code [2] and unused functions.

--
1. https://github.com/remyoudompheng/go-misc/tree/2d6ac652a50e368c874b15d00ea5e346e240e906/deadcode
2. https://en.wikipedia.org/wiki/Dead_code


* Remove unused function getCwd

Remove the unused function called `getCwd` in the fs package. This
functions was found to be unused by the deadcode static analysis
tool [1].

--
1. https://github.com/remyoudompheng/go-misc/tree/2d6ac652a50e368c874b15d00ea5e346e240e906/deadcode


* Add unconvert static analysis tool

Run unconvert [1] as a static analysis tool upon `make analysis`. This
tool checks for redundant type conversions.

--
1. https://github.com/mdempsky/unconvert


* Add goconst static analysis tool

Run goconst [1] as a static analysis tool upon `make analysis`. This
tool checks the code base for any duplicated constants, be it strings or
numbers.

The tests are ignored because the error strings for tests are often
reported as duplicates while not really being a problem.

--
1. https://github.com/jgautheron/goconst


* Add unparam static analysis tool

Run unparam [1] as a static analysis tool upon `make analysis`. This
tool checks the code base for function parameters that are unused in the
function body.

--
1. https://github.com/mvdan/unparam


* Reduce number of duplicate strings


* Add mnd static analysis tool

Run mnd (Magic Number Detector) [1] as a static analysis tool upon `make
analysis`. This tool checks the code base for magic numbers [2].

--
1. https://github.com/tommy-muehle/go-mnd
2. https://en.wikipedia.org/wiki/Magic_number_(programming)


* Define and use Makefile variables for install

* Add prealloc static analysis tool

Run prealloc [1] as a static analysis tool upon `make analysis`. This
tool checks the code base for variables that could be pre-allocated but
aren't.

--
1. https://github.com/alexkohler/prealloc


* Add nakedret, unimport, gochecknoinits as linters

Add a couple of linting tools (differentiated from "analysis" in that
they focus more, arguably, on stylistic preferences). In particular:

- nakedret [1], which is currently used to completely disallow naked
  return statements.
- unimport [2], which disallows import renaming unless it is necessary
  to avoid naming conflicts.
- gochecknoinits [3], which is used to disallow `init()` functions
  (which may cause unexpected side effects).

One other tool that has been considered but not added is
gochecknoglobals [4] which would disallow Package-scope globals. This
would further reduce side-effects. I decided not too add this as it
seems a bit too restrictive for my taste, and would require too much
refactoring at this stage of development.

--
1. https://github.com/alexkohler/nakedret
2. https://github.com/alexkohler/unimport
3. https://github.com/leighmcculloch/gochecknoinits
4. https://github.com/leighmcculloch/gochecknoglobals


* Add dogsled static analysis tool

Run dogsled [1] as a static analysis tool upon `make analysis`. This
tool checks for a "so-called" [2] go programming construct where most of
the return values of a function are ignored using blank identifiers (_).

--
1. https://github.com/alexkohler/dogsled
2. https://www.reddit.com/r/golang/comments/9syjj8/what_are_some_red_flags_for_you_in_go_code_reviews/e8sgygf/


* Add exhaustive static analysis tool

Run exhaustive [1] as a static analysis tool upon `make analysis`. This
tool checks if switch-statements on enums are exhaustive. I.e. if the
switch-statement covers all available values of the enum.

--
1. https://github.com/nishanths/exhaustive


* Add looppointer static analysis tool

Run looppointer [1] as a static analysis tool upon `make analysis`. This
tool finds pointers for loops variables which are a common cause of bugs.

I opted for looppointer [1] over exportloopref [2] as it will report
more issues. Since there are currently no issues, I think this is the best
approach to avoid introducing any bugs in the future.

--
1. https://github.com/kyoh86/looppointer
2. https://github.com/kyoh86/exportloopref


* Add gocritic static analysis tool

Run gocritic [1] as a static analysis tool upon `make analysis`. This
tool finds diagnostic, stylistic, and performance-related issues. The
tool is used with the -enableAll flag, which provides as much feedback
as possible.

In addition, this fixes the problems reported by gocritic for the
existing code base.

--
1. https://github.com/go-critic/go-critic


* Add staticcheck static analysis tool

Run staticcheck [1] as a static analysis tool upon `make analysis`. This
tool performs a number of checks of various kinds on the code base.

In addition the 4 problems reported by staticcheck have been fixed as
well.

--
1. https://staticcheck.io/docs/


* Add aligncheck static analysis tool

Run aligncheck [1] as a static analysis tool upon `make analysis`. This
tool checks if struct sizes can be improved by reordering fields.

aligncheck was chosen over maligned [1] because it has better (both
offline and online) documentation.

--
1.  https://gitlab.com/opennota/check
2. https://github.com/mdempsky/maligned


* Add structcheck and varcheck static analysis tools

Run structcheck [1] and varcheck [1] as static analysis tools upon `make
analysis`. These tools, resp., check for unused structure fields and
package global variables and constants.

--
1. https://gitlab.com/opennota/check


* Run analysis in CI

Run `make analysis` after `make lint` in the CI.

Also clean up the changes to the Makefile:
- Organize installs
- Organize analysis
- Remove gosec (leave it as CI step)
- Beautify `make lint`


* Add development dependencies (as indirect) to go.mod