Add Dockerfile
Add a Dockerfile that can be used to build an image for the development
of wordrow. The image will contain the necessary prerequisites for the
project and include a copy of the project. However, it is recommended
to use the image as a container with the local project directory
mounted.