Simplify dependency installation
Simplify the dependency (project and dev) installation with a single
command. This ensures that the normal dependencies installed using `make
install` are the correct ones (i.e. in accordance with go.mod & go.sum).

With the new Dockerfile I discovered that after running `make install`,
running `make build` (on a fresh system) would install the dependencies
again. This solves that problem.