Add Dockerfile (#121)
* Add Dockerfile

Add a Dockerfile that can be used to build an image for the development
of wordrow. The image will contain the necessary prerequisites for the
project and include a copy of the project. However, it is recommended
to use the image as a container with the local project directory
mounted.

* Simplify dependency installation

Simplify the dependency (project and dev) installation with a single
command. This ensures that the normal dependencies installed using `make
install` are the correct ones (i.e. in accordance with go.mod & go.sum).

With the new Dockerfile I discovered that after running `make install`,
running `make build` (on a fresh system) would install the dependencies
again. This solves that problem.

* Update Contributing Guidelines

Add section on using Docker for development to the Contributing
Guidelines.