Add goparamcount as analysis tool
Add the goparamcount static analysis tool on install and when running
`make analysis`. This tool will look at all function definitions in the
source code and report any with more than X parameters.

Currently, X is set to 3, as there are 4 functions that have exactly 3
parameters. On of them is in a _fuzz file, which can be safely ignored,
and another is in a testing directory, which can also be safely ignored.
However, goparamcount does not yet support custom ignore patters.

The other two functions with 3 parameters (see below) would need to be
refactored such that X can be set to 2.

- `doParseOneArgument` in package cli
- `replaceOne` in package replace