Proof of concept bufio.reader-based CSV parser
On top of a []byte-based CSV parser, this adds a bufio-based CSV parser.
With a benchmarking function for each implementation, the performance
of each looks as follows (on an 8-core i7 laptop):

- String (into map[string]string): 3532 ns/op, 22 allocs/op
- Byte array (into ByteArray): 2771 ns/op, 16 allocs/op
- Reader (into ByteArray): 2117 ns/op, 12 allocs/op

Indicating that building parser on (byte-based) bufio readers would
probably be most efficient in most use cases (given that files are read
into readers anyway).

(As an aside, this approach will probably make the MarkDown parser
easier to understand.)