Define behaviour when defining a mapping multiple times (#134)
* Log overwriting when merging maps

Add a log in the function to merge mappings in the parse/common package
to report any instance where a value already in the map is defined
again. The log is of the "debug"  kind so that it only shows up in
--verbose mode.

* Update main/wordmap to use mapping's merge func

Refactor the main package's wordmap code to use the MergeMaps function
from mappings/common  (instead of  the locally defined merge function). This
reduces code duplication and, more importantly, makes sure overwriting
mappings is also reported with inline mappings.

* Update CHANGELOG

* Add documentation on program caveats

Add a new documentation file that documents all program caveats that
apply to wordrow. This list includes the existing point found in the
"Notes" section of the `mapping-files.md` documentation file.

* Refactor /mappings/common into /common