Update static analysis tools (#133)
* Add goparamcount as analysis tool

Add the goparamcount static analysis tool on install and when running
`make analysis`. This tool will look at all function definitions in the
source code and report any with more than X parameters.

Currently, X is set to 3, as there are 4 functions that have exactly 3
parameters. On of them is in a _fuzz file, which can be safely ignored,
and another is in a testing directory, which can also be safely ignored.
However, goparamcount does not yet support custom ignore patters.


* Add gocyclo as static analysis tool

Add the gocyclo [1] static analysis tools which reports on functions
with a high cognitive complexity [2].

It is configured to report any function with a cognitive complexity of
over 15, following the example of Go Report Card [3].

--
1. https://github.com/fzipp/gocyclo
2. https://docs.codeclimate.com/docs/cognitive-complexity
3. https://goreportcard.com/report/github.com/ericcornelissen/wordrow


* Add wastedassign as static analysis tool

Add the wastedassign [1] static analysis which reports on assignments
that are wasted (i.e. variables not used after assignment).

--
1. https://github.com/sanposhiho/wastedassign


* Add shadows as static analysis tool

Add the shadows [1] static analysis tool which reports when assignments
to a variable shadow that variable name.

Shadowing variables can easily make code confusing as it is not always
clear to which instance of the variables a statement references. Hence,
it's best to avoid shadowing.

--
1. golang.org/x/tools/go/analysis/passes/shadow/cmd/shadow


* add misspell analysis tool

Add the misspell [1] analysis tool which performs a basic check for
English spelling errors in the source code.

--
1. https://github.com/client9/misspell