Fix runtime errors due to unexpected use of affix notation (#138)
* Test for corner cases for affix notation

Add a series of tests cases for corner cases with affix notation in
package replace. In addition, the bugs related to these tests have been
fixed.


* Move mapping validation to replace.go

Move the validation of mapping values from match.go to replace.go. This
enables validation of the "to" value as well as the "from" value in one
place.


* Prevent range errors due to double affix detection

e.g. replacing "- b" in "a b b".


* Update CHANGELOG