Create GitHub Actions Workflow to create Releases
Add a new GitHub Actions Workflow that will create Releases based on git
tags. Whenever a git tag is pushed starting with a "v", this Workflow
will create a new GitHub Release for it. The release message will be
taken from the git tag annotation, and release assets are automatically
generated and uploaded.

To make this possible we use actions/create-release-at-v1 [1] to create
the GitHub release, actions/upload-release-asset-at-v1 [2] to upload the
release assets, dawidd6/action-get-tag-at-v1 [3] to (easily) get the tag
name from the environment, and ericcornelissen/git-tag-annotatino-action
-at-v1 [4] to (easily) get the annotation message.

As described in the "Known issues" section of the README of [4], this
Workflow uses actions/checkout-at-v1 instead of v2 due to
incompatibility.

--
1. https://github.com/actions/create-release
2. https://github.com/actions/upload-release-asset
3. https://github.com/kceb/git-message-action
4. https://github.com/ericcornelissen/git-tag-annotation-action