Version bump
Increase the version number from 'v0.6.0-beta' to 'v0.7.0-beta' in the
program source code and the CHANGELOG. Also slightly reorder the changes
for v0.7.0-beta in the CHANGELOG.

Also, a minor fix for the SECURITY.md file. The email address in the
mailto: link contained a typo which should now be fixed.