Proof of concept byte-based replace package
Add a proof of concept byte-based replacement implemenation in the
replace package. Unfortunately, the results here are not quite in line
with the findings for byte/reader-based CSV and MD parser. In fact, this
implementation seems a bit slower (on an 8-core i7 laptop):

-  Using map[string]string: 19121 ns/op, 15 allocs/op
- Using [][]byte: 16499 ns/op, 18 allocs/op
- Using ByteMap: 17355 ns/op, 18 allocs/op

From my observations it seems that
this is due to 1) the fact that `for range` is faster than `for i` and
2) the fact that creating regexpes from []bytes is more expensive than
from strings.

NOTE: the replacement implementation has been slightly simplified in
this analysis, specifically the maintainFormatting logic has been
disabled for both map[string]string and byte-based replacement. It
may be worthwile to investigate if inclusion of this gives the bytes-
based approach an advantage.