Reader-to-map[string]string parsers
Add proof of concept reader-based parsers for both CSV and MD files that
output a map[string]string instead of a ByteMap. From this it seems that
the performance boost comes primarily from using the reader, rather than
any string conversion.

For CSVs (*):

-  Old: 3572 ns/op, 22 allocs/op
-  Reader-to-byte: 2068 ns/op, 12 allocs/op
-  Reader-to-map: 1750 ns/op, 12 allocs/op

For MarkDown (*):

-  Old: 4450 ns/op, 20 allocs/op
-  Reader-to-byte: 2690 ns/op, 7 allocs/op
-  Reader-to-map: 2941 ns/op, 9 allocs/op

For CSV, there is no noticable difference between the two new
implementations (when running multiple times either can appear faster).
For MarkDown it is clear that the conversion to strings from bytes
reduces performance, but the difference is "minor". More specifically,
it is less slower than replacing using a ByteMap (compared to
map[string]string) when replacing is.

--
(*): on an 8-core i7 laptop